+++ packages/builder/src/pages/builder/app/[application]/_layout.svelte
@@ -1,13 +1,95 @@
-<script lang="ts">
-  import { featureFlags } from "@/stores/portal"
-  import NewLayout from "./_flagged/layout.new.svelte"
-  import OldLayout from "./_flagged/layout.old.svelte"
+<script>
+  import {
+    initialise,
+    reset,
+    builderStore,
+    previewStore,
+    deploymentStore,
+  } from "@/stores/builder"
+  import { appsStore } from "@/stores/portal"
+  import { notifications } from "@budibase/bbui"
+  import { API } from "@/api"
+  import { redirect } from "@roxi/routify"
+  import { onDestroy } from "svelte"
+  import BuilderSidePanel from "./_components/BuilderSidePanel.svelte"
+  import PreviewOverlay from "./_components/PreviewOverlay.svelte"
+  import EnterpriseBasicTrialModal from "@/components/portal/onboarding/EnterpriseBasicTrialModal.svelte"
+  import SideNav from "./_components/SideNav/SideNav.svelte"
 
-  export let application: string
+  export let application
 
-  $: layout = $featureFlags.WORKSPACES ? NewLayout : OldLayout
+  let promise = getPackage()
+
+  async function getPackage() {
+    try {
+      reset()
+      const pkg = await API.fetchAppPackage(application)
+      await initialise(pkg)
+      await appsStore.load()
+      await deploymentStore.load()
+      return pkg
+    } catch (error) {
+      notifications.error(`Error initialising app: ${error?.message}`)
+      $redirect("../../")
+    }
+  }
+
+  onDestroy(() => {
+    // Run async on a slight delay to let other cleanup logic run without
+    // being confused by the store wiping
+    setTimeout(reset, 10)
+  })
 </script>
 
-<svelte:component this={layout} {application}>
-  <slot />
-</svelte:component>
+{#if $builderStore.builderSidePanel}
+  <BuilderSidePanel />
+{/if}
+
+<div class="root" class:blur={$previewStore.showPreview}>
+  <SideNav />
+  {#await promise}
+    <!-- This should probably be some kind of loading state? -->
+    <div class="loading" />
+  {:then _}
+    <div class="body">
+      <slot />
+    </div>
+  {:catch error}
+    <p>Something went wrong: {error.message}</p>
+  {/await}
+</div>
+
+{#if $previewStore.showPreview}
+  <PreviewOverlay />
+{/if}
+
+<EnterpriseBasicTrialModal />
+
+<style>
+  .loading {
+    min-height: 100%;
+    height: 100%;
+    width: 100%;
+    background: var(--background);
+  }
+  .root {
+    min-height: 100%;
+    height: 100%;
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+    transition: filter 260ms ease-out;
+    align-items: stretch;
+    position: relative;
+  }
+  .root.blur {
+    filter: blur(8px);
+  }
+  .body {
+    flex: 1 1 auto;
+    width: 0;
+    z-index: 1;
+    display: flex;
+    flex-direction: column;
+  }
+</style>
